refactor: #149 - 팔로우, 언팔로우 fetch 함수 파일명 수정
