+++ src/components/common/FollowList/FollowList.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, Fragment, SetStateAction } from 'react'
 import useFollowQuery from '@/components/common/FollowList/hooks/useFollowQuery'
 import useInfiniteScroll from '@/hooks/useInfiniteScroll'
-import { FetchGetFollowProps } from '@/services/user/follow/route'
+import { FetchGetFollowProps } from '@/services/user/follow/follow'
 import User from '../User/User'
 
 export interface FollowListProps {
